requirements.txt
Pillow==6.1.0
requests==2.22.0
urllib3==1.25.3